fix(server): float32 everywhere on the full matrix, locked and purged viz caches, recs in the global subsets, cluster-wide byte size
rank.scores/centrality cast to float32 rather than float64, so the full
corpus matrix is no longer promoted (and doubled) on every cosine request;
/viz/histogram routes through _similarity to reuse the cached unit matrix.

One _VIZ_CACHE_LOCK now guards every read-modify-write of the subset, top8
and MST caches, and every subset miss purges the entries belonging to a
superseded full matrix -- a stale key otherwise pinned a whole ~450 MB
corpus matrix that _MATRIX_CACHE had already replaced. _UNIT_CACHE holds
its matrix instead of a bare id() for the same reason. Keep counts 2 -> 4.

/viz/tour, /viz/mst, /viz/hubs and /viz/extremes take an optional `recs`
list so a `surprise` rec outside the seed's nearest ring still has a row
to highlight; /viz/map passes the corpus pair it already holds; walk and
MST stop re-casting the subset, which restores the shared pairwise cache.

store.data_size_bytes sums dbStats across the whole cluster (Atlas bills
the cluster, not one database) with two documented fallbacks.

+++ src/music_recommendations/server/rank.py
@@ -33,8 +33,11 @@ def scores(seed: np.ndarray, matrix: np.ndarray, metric: str = "cosine") -> np.n
     "euclidean" -> (0, 1], 1/(1+distance), so it is still a similarity and
                    sorts the same direction as cosine.
     """
-    matrix = np.asarray(matrix, dtype=float)
-    seed = np.asarray(seed, dtype=float)
+    # float32, not float64: the full corpus matrix is stored float32, and
+    # promoting it here doubled every temporary — ~900 MB of allocation per
+    # cosine call at 90k tracks, for precision no similarity score needs.
+    matrix = np.asarray(matrix, dtype=np.float32)
+    seed = np.asarray(seed, dtype=np.float32)
     if metric == "euclidean":
         return 1.0 / (1.0 + np.linalg.norm(matrix - seed, axis=-1))
     if metric != "cosine":
@@ -54,7 +57,7 @@ def centrality(matrix: np.ndarray, metric: str = "cosine") -> np.ndarray:
     O(n*d) rather than the O(n^2) full similarity matrix. That matters: at
     10,000 tracks the naive matrix is 0.4 GB, this is 40 KB.
     """
-    matrix = np.asarray(matrix, dtype=float)
+    matrix = np.asarray(matrix, dtype=np.float32)
     if metric == "cosine":
         unit = _normalize(matrix)
         return unit @ unit.mean(axis=0)
